+++ src/math/matrix3.js
@@ -299,38 +299,48 @@ class Matrix3d {
     }
 
     /**
-     * apply the current transform to the given 3d vector
+     * apply the current transform to the given 2d or 3d vector
      * @name apply
      * @memberOf me.Matrix3d
      * @function
-     * @param {me.Vector3d} vector the vector object to be transformed
-     * @return {me.Vector3d} result vector object.
+     * @param {me.Vector2d|me.Vector3d} vector the vector object to be transformed
+     * @return {me.Vector2d|me.Vector3d} result vector object.
      */
      apply(v) {
         var a = this.val,
         x = v.x,
         y = v.y,
-        z = v.z;
+        z = (typeof v.z !== "undefined") ? v.z : 1;
 
         var w = (a[3] * x + a[7] * y + a[11] * z + a[15]) || 1.0;
 
         v.x = (a[0] * x + a[4] * y + a[8] * z + a[12]) / w;
         v.y = (a[1] * x + a[5] * y + a[9] * z + a[13]) / w;
-        v.z = (a[2] * x + a[6] * y + a[10] * z + a[14]) / w;
+
+        if (typeof v.z !== "undefined") {
+            v.z = (a[2] * x + a[6] * y + a[10] * z + a[14]) / w;
+        }
 
         return v;
      }
 
      /**
-      * apply the inverted current transform to the given 3d vector
+      * apply the inverted current transform to the given 2d or 3d vector
       * @name applyInverse
       * @memberOf me.Matrix3d
       * @function
-      * @param {me.Vector3d} vector the vector object to be transformed
-      * @return {me.Vector3d} result vector object.
+      * @param {me.Vector2d|me.Vector3d} vector the vector object to be transformed
+      * @return {me.Vector2d|me.Vector3d} result vector object.
       */
      applyInverse(v) {
-         // XXX : TODO
+         // invert the current matrix
+         var im = pool.pull("Matrix3d", this).invert();
+         
+         // apply the inverted matrix
+         im.apply(v);
+
+         pool.push(im);
+
          return v;
      }