+++ tests/spec/mat3d-spec.js
@@ -95,19 +95,19 @@ describe("me.Matrix3d", function () {
         expect(matA.toString() === result).toEqual(true);
     });
 
-    xit("should multiply a 2d vector properly", function () {
+    it("should multiply a 2d vector properly", function () {
         var matA = new me.Matrix3d(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1);
         var vecA = new me.Vector2d(1, 2);
 
         matA.apply(vecA);
         //multiply back with the inverted matrix
-        matA.invert().apply(vecA);
+        matA.applyInverse(vecA);
 
         // and we should have back the original vector values
         expect(vecA.toString()).toEqual("x:1,y:2");
     });
 
-    xit("should multiply a 3d vector properly with the inverted matrix", function () {
+    it("should multiply a 3d vector properly with the inverted matrix", function () {
         var matA = new me.Matrix3d(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 2, 3, 1);
         var vecA = new me.Vector3d(3, 7, 1);