+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Container : Containers defines an additional `enableChildBoundsUpdate` flag to enable full bounds update, including child bounds (disabled by default)
 - Entity : me.CollectableEntity is now deprecated and replaced by a more generic `me.Collectable` base object that do not extend me.Entity anymore
 - Entity : me.LevelEntity is now deprecated and replaced by a more generic `me.Trigger` objects that do not extend me.Entity anymore
+- Math : add missing `applyInverse()` method to the Matrix3d implementation
 - Renderable : fix Renderables bounds not correctly respecting the anchor point.
 - Renderer : WebGL2 is now the default mode when using the WebGL renderer (use `preferWebGL1 = true` calling me.video.init if you need to force WebGL1)
 - Tiled : add TMX `tintcolor` parsing for tile and object layers